Basic UI For The LG Gesture And Voice Control App.
Implemented the following features and changes in the Liquid Galaxy gesture and voice control app:

Landing Page:

Developed a basic user interface for the landing page, providing users with the first interaction point of the application.

Added two prominent buttons on the landing page, allowing users to access the camera and voice control functionalities.

Incorporated a floating button on the landing page to provide easy access to the main menu.

Implemented an icon button on the app bar, enabling users to access the help menu.

Help Menu:

Created a visually appealing help menu within the application, serving as a reference point for users to understand available commands.

Implemented a ListView component to display a list of commands supported by the LG Gesture and voice control app.

Designed the help menu to showcase the logo of the Liquid Galaxy project, providing a consistent visual identity.

Menu:

Started developing the menu section of the application, focusing on providing users with additional features and options.

Ongoing work in progress, aiming to create a comprehensive menu that enhances the usability and functionality of the application.

By incorporating these elements into the application, I have improved the user experience and made it more intuitive for users to interact with the Liquid Galaxy gesture and voice control app.